--- .dockerignore
@@ -1,7 +1,0 @@
-assets/
-.vscode/
-bytegolf
-vendor/
-frontend/
-scripts/
-terraform/
