+++ .gitignore
@@ -1,6 +1,6 @@
 .DS_STORE
 .vscode/
-bytegolf
+bytegolf/bytegolf
 *.exe
 *.exe~
 subs/
