+++ frontend/static/assets/golf-ball/ball.psd
