+++ frontend/static/assets/golf-ball/loading-icon/loading-icon.aep
