+++ go.mod
@@ -5,16 +5,12 @@ require (
 	cloud.google.com/go/pubsub v1.3.1
 	github.com/Squwid/go-randomizer v0.1.2
 	github.com/golang-jwt/jwt v3.2.2+incompatible
-	github.com/google/uuid v1.2.0
 	github.com/gorilla/mux v1.7.4
-	github.com/mitchellh/mapstructure v1.4.1
 	github.com/sirupsen/logrus v1.8.1
 	github.com/uptrace/bun v1.1.9
 	github.com/uptrace/bun/dialect/pgdialect v1.1.9
 	github.com/uptrace/bun/driver/pgdriver v1.1.9
 	github.com/uptrace/bun/extra/bundebug v1.1.9
-	google.golang.org/api v0.21.0
-	google.golang.org/grpc v1.38.0
 )
 
 require (
@@ -42,8 +38,10 @@ require (
 	golang.org/x/sys v0.3.0 // indirect
 	golang.org/x/text v0.5.0 // indirect
 	golang.org/x/tools v0.1.12 // indirect
+	google.golang.org/api v0.21.0 // indirect
 	google.golang.org/appengine v1.6.5 // indirect
 	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	google.golang.org/grpc v1.38.0 // indirect
 	google.golang.org/protobuf v1.25.0 // indirect
 	honnef.co/go/tools v0.0.1-2020.1.3 // indirect
 	mellium.im/sasl v0.3.0 // indirect
