+++ lib/api/random.go
