+++ lib/auth/auth.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/Squwid/bytegolf/sqldb"
+	"github.com/Squwid/bytegolf/lib/sqldb"
 	"github.com/golang-jwt/jwt"
 	"github.com/sirupsen/logrus"
 )
