+++ lib/globals/env.go
