+++ lib/globals/languages.go
