+++ tests/populate_tables_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Squwid/bytegolf/api"
-	"github.com/Squwid/bytegolf/sqldb"
+	"github.com/Squwid/bytegolf/lib/api"
+	"github.com/Squwid/bytegolf/lib/sqldb"
 	"github.com/Squwid/go-randomizer"
 	"github.com/sirupsen/logrus"
 )
